fixed codacy issues
