--- static/scripts/helpers/cookieHelper.js
@@ -1,8 +1,0 @@
-export function getCookie(name) {
-  	const value = `; ${document.cookie}`;
-  	const parts = value.split(`; ${name}=`);
-  	if (parts.length === 2) {
-  		return parts.pop().split(';').shift();
-  	}
-  	return undefined;
-}
